zrq's second homework
