+++ The first Homework/张若琦_2017201981/ExtractURL.h
@@ -1,0 +1,10 @@
+#ifndef ExtractURL_H_
+#define ExtractURL_H_
+
+using std :: string;
+
+void ExtractURL(const string & root_url,const string & now_url,const string &,const string & origin_url);
+
+void Getfile_and_ExtractURL(const string & root_url,const string & now_url,string & str_file);
+
+#endif // ExtractURL_H_
