+++ The first Homework/张若琦_2017201981/Main.cpp
@@ -1,10 +1,15 @@
 #include <string>
 #include <iostream>
 #include "crawl.h"
+#include "clean.h"
 
-int main()
+int main(int argc,char * argv[])
 {
-	string root_url = "info.ruc.edu.cn";
-	crawl(root_url);
+	string root_url = argv[1];
+	if (root_url == "clean")
+		clean();
+	else
+		crawl(root_url);
+	
 	return 0;
 }
