+++ The first Homework/张若琦_2017201981/Makefile
@@ -1,7 +1,10 @@
-objects = Main.o crawl.o
+objects = Main.o crawl.o ExtractURL.o ReadFile.o DisposeStr.o clean.o
 
 crawl : $(objects)
-	g++ -o crawl $(objects)
+	g++ -std=c++11 -o crawl $(objects)
+
+DisposeStr.o : DisposeStr.cc DisposeStr.h
+	g++ -std=c++11 -c DisposeStr.cc
 
 .PHONY : clean
 clean :
