+++ The first Homework/张若琦_2017201981/ReadFile.h
@@ -1,0 +1,10 @@
+#ifndef ReadFile_H_
+#define ReadFile_H_
+
+using std :: string;
+
+static const int MAXN = 1000;
+
+string get_file(const char* filename);
+
+#endif // ReadFile_H_
