+++ The first Homework/张若琦_2017201981/clean.cc
@@ -1,0 +1,9 @@
+#include <cstdlib>
+
+void clean()
+{
+	system("rm *.html");
+	system("rm *.php");
+	system("rm list.txt");
+	system("rm tmp");
+}
