+++ The first Homework/张若琦_2017201981/clean.h
@@ -1,0 +1,6 @@
+#ifndef clean_H_
+#define clean_H_
+
+void clean();
+
+#endif // clean_H_
