+++ The first Homework/张若琦_2017201981/crawl.h
@@ -1,43 +1,18 @@
-#ifndef __MYCRAWL__
-#define __MYCRAWL__
+#ifndef CRAWL_H_
+#define CRAWL_H_
 
-#define MAXN 1000
-#define MAX_URL_NUM 3000
-#define ROOT_URL_LEN 15
-
-using std :: cin;
-using std :: cout;
-using std :: endl;
 using std :: string;
 
-bool check_out_site(string url,string root_url);
+static const int MAXN = 1000;
 
-bool check_in_set(string url);
+void crawl(const string & root_url);
 
-bool check_php(string str);
-
-bool check_html(string str);
-
-bool check_css(string str);
-
-bool check_js(string str);
-
-bool check_suf(string str);
-
-string check_pre(string url);
-
-string filter_blank_character(string url);
 
 void Delay();
 
-string get_file(const char* filename);
+inline void Delay()
+{
+	for (int i = 1;i <= 10000000;i ++);
+}
 
-string Get_whole_url(string url,string root_url,string now_url);
-
-void ExtractURL(string root_url,string now_url,string origin_url);
-
-string Getfile_and_ExtractURL(string root_url,string now_url);
-
-void crawl(string root_url);
-
-#endif // __MYCRAWL__
+#endif // CRAWL_H_
